Add camera movement and correct tons of vector orientation mistakes

+++ src/util.rs
@@ -1,7 +1,7 @@
-use std::f32::consts::{PI, TAU};
+use std::f64::consts::{PI, TAU};
 
 use cgmath::Rad;
 
-pub fn normalize_angle(a: Rad<f32>) -> Rad<f32> {
-    Rad(a.0 - (TAU * f32::floor((a.0 + PI) / TAU)))
+pub fn normalize_angle(a: Rad<f64>) -> Rad<f64> {
+    Rad(a.0 - (TAU * f64::floor((a.0 + PI) / TAU)))
 }
